Tesoreria: percepcion en dashboard, fix cierre de caja y tab de inventario
- Card de percepcion acumulada en dashboard (CUENTAS), sumada desde Compra
- Fix: cierre de caja ya no retira/zera cuentas digitales (Plin, Yape, etc.),
  el retiro fisico solo aplica a EFECTIVO
- Permite varios cierres de caja por cuenta el mismo dia para reconciliar
  con mas frecuencia (se quita la constraint unica de Cierre_Diario)
- Nuevo tab "Inventario" en Tesoreria: productos agrupados por categoria
  con costo invertido y valor de mercaderia
- DbMigrationRunner: corrige constraint de Cierre_Diario via JdbcTemplate
  (data.sql no soporta bloques DO $$ de Postgres, Spring corta por ";")

+++ BACKEND/src/main/java/com/AdrithStore/backend/repository/ProductoRepository.java
@@ -4,6 +4,7 @@ import com.AdrithStore.backend.model.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -33,7 +34,7 @@ public interface ProductoRepository extends JpaRepository<Producto, Integer> {
     List<Producto> findStockBajo();
 
     
-    List<Producto> findByStockLessThan(int cantidad);
+    List<Producto> findByStockLessThan(BigDecimal cantidad);
